Add EditCoInsuredShared
