+++ Projects/App/Sources/Journeys/CoInsuredJourney.swift
@@ -1,5 +1,6 @@
 import Contracts
 import EditCoInsured
+import EditCoInsuredShared
 import Presentation
 import hCore
