+++ Projects/Contracts/Sources/Models/ContractModels.swift
@@ -1,3 +1,4 @@
+import EditCoInsuredShared
 import Foundation
 import Presentation
 import hCore
