+++ Projects/Contracts/Sources/Stores/ContractAction.swift
@@ -1,4 +1,5 @@
 import Apollo
+import EditCoInsuredShared
 import Presentation
 import SwiftUI
 import TerminateContracts
