+++ Projects/Contracts/Sources/Stores/ContractState.swift
@@ -1,8 +1,7 @@
 import Apollo
+import EditCoInsuredShared
 import Presentation
 import SwiftUI
-import hCore
-import hGraphQL
 
 public struct ContractState: StateProtocol {
