+++ Projects/EditCoInsured/Sources/Services/DemoImplementation/EditCoInsuredDemoClient.swift
@@ -1,3 +1,4 @@
+import EditCoInsuredShared
 import Foundation
 import hCore
 import hGraphQL
