+++ Projects/EditCoInsured/Sources/Services/OctopusImplementation/EditCoInsuredServiceOctopus.swift
@@ -1,3 +1,4 @@
+import EditCoInsuredShared
 import Foundation
 import hCore
 import hGraphQL
