+++ Projects/EditCoInsured/Sources/Services/Protocols/EditCoInsuredService.swift
@@ -1,3 +1,4 @@
+import EditCoInsuredShared
 import Foundation
 import hCore
