+++ Projects/EditCoInsured/Sources/View/CoInsuredSelectScreen.swift
@@ -1,3 +1,4 @@
+import EditCoInsuredShared
 import Presentation
 import SwiftUI
 import hCore
