+++ Projects/EditCoInsured/Sources/View/CoInusuredInput.swift
@@ -1,4 +1,5 @@
 import Combine
+import EditCoInsuredShared
 import Presentation
 import SwiftUI
 import hCore
