+++ Projects/EditCoInsured/Sources/View/NewCoInsured/InsuredPeopleNewScreen.swift
@@ -1,3 +1,4 @@
+import EditCoInsuredShared
 import Presentation
 import SwiftUI
 import hCore
