+++ Projects/EditCoInsured/Sources/View/NewCoInsured/RemoveCoInsuredScreen.swift
@@ -1,3 +1,4 @@
+import EditCoInsuredShared
 import Foundation
 import Presentation
 import SwiftUI
