+++ Projects/EditCoInsuredShared/Sources/Views/ContractOwnerField.swift
@@ -1,0 +1,71 @@
+import SwiftUI
+import hCoreUI
+
+public struct ContractOwnerField: View {
+    let enabled: Bool?
+    let hasContentBelow: Bool
+    let fullName: String
+    let SSN: String
+
+    public init(
+        enabled: Bool? = false,
+        hasContentBelow: Bool,
+        fullName: String,
+        SSN: String
+    ) {
+        self.enabled = enabled
+        self.hasContentBelow = hasContentBelow
+        self.fullName = fullName
+        self.SSN = SSN.displayFormatSSN ?? ""
+    }
+
+    public init(
+        enabled: Bool? = false,
+        hasContentBelow: Bool,
+        config: InsuredPeopleConfig
+    ) {
+        self.enabled = enabled
+        self.hasContentBelow = hasContentBelow
+        self.fullName = config.holderFullName
+        self.SSN = config.holderSSN?.displayFormatSSN ?? ""
+    }
+
+    public var body: some View {
+        VStack {
+            VStack(alignment: .leading, spacing: 0) {
+                HStack {
+                    hText(fullName)
+                        .foregroundColor(getTitleColor)
+                    Spacer()
+                    Image(uiImage: hCoreUIAssets.lockSmall.image)
+                        .foregroundColor(hTextColor.tertiary)
+                }
+                hText(SSN, style: .footnote)
+                    .foregroundColor(getSubTitleColor)
+            }
+            .frame(maxWidth: .infinity, alignment: .leading)
+            if hasContentBelow {
+                Divider()
+            }
+        }
+        .padding(.bottom, hasContentBelow ? 0 : 16)
+    }
+
+    @hColorBuilder
+    var getTitleColor: some hColor {
+        if enabled ?? false {
+            hTextColor.primary
+        } else {
+            hTextColor.tertiary
+        }
+    }
+
+    @hColorBuilder
+    var getSubTitleColor: some hColor {
+        if enabled ?? false {
+            hTextColor.secondary
+        } else {
+            hTextColor.tertiary
+        }
+    }
+}
