+++ Projects/TravelCertificate/Project.swift
@@ -4,7 +4,7 @@ import ProjectDescriptionHelpers
 let project = Project.framework(
     name: "TravelCertificate",
     targets: Set([.framework]),
-    projects: ["hCore", "hCoreUI", "Contracts"],
+    projects: ["hCore", "hCoreUI", "Contracts", "EditCoInsuredShared"],
     sdks: [],
     includesGraphQL: false
 )
