+++ Projects/TravelCertificate/Sources/Views/WhoIsTravelingScreen.swift
@@ -1,4 +1,5 @@
 import Contracts
+import EditCoInsuredShared
 import Foundation
 import Presentation
 import SwiftUI
